refactor(convertId.ts): moving convertId.ts
